GNI2014 {treemap}	R Documentation

GNI 2014 Data

Description

Gross national income (per capita) in dollars and population totals per country in 2014.

Details

The GNI numbers from the World Bank are based on the Atlas. The population data are taken from Natural Earth Data.

References

The World Bank - GNI per capita ranking Natural Earth Data

[Package treemap version 2.4-3 Index]