business {treemap}	R Documentation

Fictitious Business Statistics Data

Description

Fictitious (aggregated) business statistics data. The index variables (NACE1 to NACE4) are derived from the Statistical Classification of Economic Activities in the European Community (NACE). The variables turnover(.prev) and employees(.prev) have values for NACE codes in the business economy domain only.

References

Statistical Classification of Economic Activities in the European Community (NACE) Eurostat - Structural business statistics (SBS)

[Package treemap version 2.4-3 Index]