itreemap {treemap}	R Documentation

Interactive user interface for treemap

Description

This function is an interactive user interface for creating treemaps. Interaction is provided for the four main input arguments of (treemap) besides the data.frame itself, namely index, vSize, vColor and type. Zooming in and out is possible. Command line outputs are generated in the console.

Usage

itreemap(
  dtf = NULL,
  index = NULL,
  vSize = NULL,
  vColor = NULL,
  type = NULL,
  height = 700,
  command.line.output = TRUE
)

Arguments

dtf
a data.frame (treemap) If not provided, then the first data.frame in the global workspace is loaded.

index
index variables (up to four). See treemap.

vSize
name of the variable that determine the rectangle sizes.

vColor
name of the variable that determine the rectangle colors. See treemap.

type
treemap type. See treemap.

height
height of the plotted treemap in pixels. Tip: decrease this number if the treemap doesn't fit conveniently.

command.line.output
if TRUE, the command line output of the generated treemaps are provided in the console.

Note

This interface will no longer be maintained (except for small bugs), since there is a better interactive interface available: https://github.com/d3treeR/d3treeR.

Examples

## Not run: 
data(business)
itreemap(business)

## End(Not run)

[Package treemap version 2.4-3 Index]