treecolors {treemap}	R Documentation

Interactive tool to experiment with Tree Colors

Description

Tree Colors are color palettes for tree data structures. They are used in treemap by default (type="index"). With this tool, users can experiment with the parameters (in treemap stored in palette.HCL.options). Tree Colors can directly be obtained by treepalette with method="HCL".

Usage

treecolors(height = 700)

Arguments

height
height of the plotted treemap in pixels. Tip: decrease this number if the treemap doesn't fit conveniently.

Examples

## Not run: 
treecolors()

## End(Not run)

[Package treemap version 2.4-3 Index]