treemap-package {treemap}	R Documentation

Treemap package

Description

A treemap is a space-filling visualization of hierarchical structures. This package offers great flexibility to draw treemaps.

Details

The main function is treemap. See also itreemap for a graphical user interface to create treemaps. By default Tree Colors are used, which are colors from the HCL color space. Use treecolors to experiment with the parameter settings.

Author(s)

Martijn Tennekes mtennekes@gmail.com

[Package treemap version 2.4-3 Index]